+++ FileHelpers.Examples/Examples/60.Sorting/10.SortBigFiles.cs
@@ -12,18 +12,27 @@ namespace ExamplesFx
     public class BigFileSort
         : ExampleBase
     {
-        //-> File:SortingWithRecord.cs
+
+        //-> If you need to sort a really big file (20Mb and more) you have the BigFileSorter
+
+        //-> Implements <a href="http://en.wikipedia.org/wiki/External_sorting">External Sorting (wikipedia)</a>
+
+        //-> The Sorter will split the file in blocks, write them to temp files, and finally join all in a sorted file
+
         public override void Run()
         {
-            // Implements http://en.wikipedia.org/wiki/External_sorting
-            // OrdersTab is IComparable<OrdersTab>
+            //-> File:SortingWithRecord.cs
 
-            var sorter = new BigFileSorter<OrdersTab>(10*1024*1024); // 10 Mb
-            sorter.DeleteTempFiles = true;
+            // OrdersTab must be IComparable<OrdersTab>
+            
+            // We recommend to split in blocks between 1 and 40 Mb
+            var sorter = new BigFileSorter<OrdersTab>(10*1024*1024); // 10 Mb blocks
+
             sorter.Sort("unsorted.txt", "sorted.txt");
+
+            //-> /File
         }
 
-        //-> /File
 
 
         //-> File:OrdersTab.cs
