+++ FileHelpers.Examples/Examples/60.Sorting/30.SortBigFilesString2.cs
@@ -12,20 +12,24 @@ namespace ExamplesFx
     public class BigFileSortString2
         : ExampleBase
     {
-        //-> File:SortingWithoutRecord.cs
+        //-> Implements <a href="http://en.wikipedia.org/wiki/External_sorting">External Sorting (wikipedia)</a>
+
+        //-> You can use any sort you want, for example sorting by line length
+
         public override void Run()
         {
-            // Implements http://en.wikipedia.org/wiki/External_sorting
+            //-> File:SortingWithoutRecord.cs
 
             var sorter = new BigFileSorter(
                 (x, y) => {
                     // You can add here any custom function
                     return x.Length.CompareTo(y.Length);
                 });
-            sorter.DeleteTempFiles = true;
+
             sorter.Sort("unsorted.txt", "sorted.txt");
+
+            //-> /File
         }
 
-        //-> /File
     }
 }
