+++ FileHelpers/Sort/BigFileSorter.T.cs
@@ -30,7 +30,7 @@ namespace FileHelpers
         /// <summary>
         /// Maximum block size,  50 meg
         /// </summary>
-        private const int MaxBufferSize = 50*1024*1024;
+        private const int MaxBufferSize = 40*1024*1024;
 
         /// <summary>
         /// Minimum block size 0.5 meg
